Add some of the base of when a new header comes in

+++ Liviano/Electrum/ElectrumClient.cs
@@ -408,9 +408,9 @@ namespace Liviano.Electrum
             return await RequestInternal<BlockchainBlockHeaderWithCheckpointHeightResult>(json);
         }
 
-        public async Task<BlockchainBlockHeadersResult> BlockchainBlockHeaders(int height, int count = 2016)
+        public async Task<BlockchainBlockHeadersResult> BlockchainBlockHeaders(long height, long count = 2016)
         {
-            var obj = new Request { Id = ++RequestId, Method = "blockchain.block.headers", Params = new List<int> { height, count, 0 } };
+            var obj = new Request { Id = ++RequestId, Method = "blockchain.block.headers", Params = new List<long> { height, count, 0 } };
             var json = Serialize(obj);
 
             return await RequestInternal<BlockchainBlockHeadersResult>(json);
